requirements.txt
streamlit
streamlit_chat
openai
load_dotenv
watchdog